add box_diff and prob_loss gpu kernel

+++ oneflow/core/kernel/normal_model_update_kernel.h
@@ -21,9 +21,9 @@ class NormalMdUpdateKernel : public KernelIf<device_type> {
                            std::function<Blob*(const std::string&)> BnInOp2Blob) const = 0;
 };
 
-bool TriggerWarmup(const NormalModelUpdateOpUserConf& conf, double lr, int64_t cur_batch_num);
-double GetWarmupLearningRate(const WarmupConf&, double lr, int64_t cur_batch_num);
-double GetDecayedLearningRate(const LearningRateDecayConf&, double lr, int64_t cur_batch_num);
+bool TriggerWarmup(const NormalModelUpdateOpUserConf& conf, float lr, int64_t cur_batch_num);
+float GetWarmupLearningRate(const WarmupConf&, float lr, int64_t cur_batch_num);
+float GetDecayedLearningRate(const LearningRateDecayConf&, float lr, int64_t cur_batch_num);
 
 #define DECLARE_MDUPDT_KERNEL_CREATOR(x) Kernel* Create##x##MdUpdtKernel(const KernelConf&);
